fix: Revert ThemeProviderLegacy rename and update new ThemeProvider to CustomPropertiesProvider
