+++ src/components/Frame/components/Toast/Toast.tsx
@@ -4,7 +4,7 @@ import {MobileCancelMajor} from '@shopify/polaris-icons';
 import {classNames} from '../../../../utilities/css';
 import {Key} from '../../../../types';
 import {Button} from '../../../Button';
-import {ThemeProviderLegacy} from '../../../ThemeProviderLegacy';
+import {ThemeProvider} from '../../../ThemeProvider';
 import {Icon} from '../../../Icon';
 import {KeypressListener} from '../../../KeypressListener';
 import type {ToastProps} from '../../../../utilities/frame';
@@ -64,13 +64,13 @@ export function Toast({
   const className = classNames(styles.Toast, error && styles.error);
 
   return (
-    <ThemeProviderLegacy theme={{colorScheme: 'inverse'}}>
+    <ThemeProvider theme={{colorScheme: 'inverse'}}>
       <div className={className}>
         <KeypressListener keyCode={Key.Escape} handler={onDismiss} />
         {content}
         {actionMarkup}
         {dismissMarkup}
       </div>
-    </ThemeProviderLegacy>
+    </ThemeProvider>
   );
 }
