--- src/components/ThemeProviderLegacy/index.ts
@@ -1,1 +1,0 @@
-export * from './ThemeProviderLegacy';
